Split out `match_as_ref`
